+++ book.rb
@@ -1,8 +1,7 @@
-require_relative 'app'
+# require_relative 'app'
 class Book
   attr_accessor :name, :author, :rentals, :title
 
-  # attr_reader :id
   def initialize(title, author)
     @title = title
     @author = author
@@ -10,7 +9,7 @@ class Book
     @id = Random.rand(1..1000)
   end
 
-  def add_rental(person, date)
-    Rental.new(date, self, person)
+  def add_rental(person, date, rental_manager)
+    rental_manager.create_rental(date, self, person)
   end
 end
