+++ person.rb
@@ -1,5 +1,7 @@
 require 'securerandom'
 require_relative 'nameable'
+# require_relative 'app'
+
 class Person < Nameable
   attr_accessor :name, :age, :rentals, :parent_permission
   attr_reader :id
@@ -17,8 +19,8 @@ class Person < Nameable
     @name
   end
 
-  def add_rental(date, book)
-    Rental.new(date, book, self)
+  def add_rental(date, book, rental_manager)
+    rental_manager.create_rental(date, book, self)
   end
 
   private
